Made fun getLogObservation for regression
